requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0

# Database and persistence
psycopg2-binary>=2.9.0
langgraph-checkpoint-postgres>=2.0.0

# Search and retrieval
tavily-python>=0.5.0

# Vector store and embeddings
faiss-cpu>=1.8.0
langchain-openai>=0.2.0  # For embeddings

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.17.0
scipy>=1.11.0

# Web scraping and content extraction
beautifulsoup4>=4.12.0
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and monitoring
structlog>=23.2.0

# Type hints
typing-extensions>=4.8.0